Verificacao de entradas validas #28 (#50)
* Padronização do controller para tratamento de erros

-Padronizacao do ClientController:

-Objetos não sao instanciados dentro do controller(Client)

-Todas as excecoes conhecidas sao tratadas no service e enviadas ao controller como excecao personalizada: ClientServiceException

* Padronização do CryptoController

-Padronizacao do CryptoController:

-Objetos não sao instanciados dentro do controller(Crypto)

-Todas as excecoes conhecidas sao tratadas no service e enviadas ao controller como excecao personalizada: CryptoServiceException

* Verificacao de entradas

ClientRepository:
-Verificar se um cliente ja esta cadastrado com o mesmo ID
-Verificar se um cliente existe ao tentar remover

CryptoRepository:
-Verifica se a criptomoeda ja foi cadastrada
-Verifica se a cripto existe ao tentar remover

PortfolioRepository:
-Remoção de metodo repetitivo findPortfolio (find e load: ambos procuram e retornam um portfolio)

ClientService e CryptoService
-Um catch para todas as possiveis excecoes lançadas por CryptoRepository e ClientRepository

* userID e portfolioID case

sempre maiusculo

* Ajustes

* Validacao de entradas - 2

PortfolioController
-Removi instanciacao de objetos dentro do controller

ReportController
-Removi instanciacao de objetos dentro do controller

CryptoCurrency
-Novo parametro: investorsAmount
ajuda na verificacao na hora de remover uma cripto: Somente remover se ninguem possuir a cripto no portfolio

Portfolio
-UserID e PortfolioID são case Sensitive

ClientRepository
-Ao remover um cliente, remover tambem seu portfolio

CryptoRepository
-Adequacao do arquivo para receber o novo campo (investorsAmount)
-Atualizacao do updateCrypto para nao depender do delete

PortfolioRepository
-Novos métodos:
-deletePortfolio: usado para deletar portfolio ao deletar cliente
-portfolioExists: verificacao se um portfolio com o mesmo userID e portfolioID já existe, impedindo duplicatas
-portfolioHasInvestments: verifica se um portfolio tem investimentos, usado no delete portfolio para impedir que portfolios com investimentos ativos sejam apagados

PortfolioService
-Atualizar quantidade de investidores nas criptos nos metodos de compra e venda

* Ajuste de erros

-erro de merge
-Apos alguns testes, lembrei de atualizar tambem o criptoConverter com a logica de investorsAmount
-atualizar cliente nao depende de deleteClient

* Verificação-de-entradas-validas

Retirada de Warnings atestados pela IDE

---------

+++ src/main/java/com/cryptomanager/repositories/CryptoRepository.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Repository;
 import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
+
 
 @Repository
 public class CryptoRepository {
@@ -13,8 +15,9 @@ public class CryptoRepository {
     private static final String FILE_PATH = "cryptos.txt";
 
     public void saveCrypto(CryptoCurrency crypto) throws IOException {
+        if(cryptoExists(crypto.getName())) { throw new IllegalArgumentException("Criptomoeda ja foi cadastrada"); }
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(FILE_PATH, true))) {
-            writer.write(crypto.toString() + "\n");
+            writer.write(crypto + "\n");
         }
     }
 
@@ -24,13 +27,17 @@ public class CryptoRepository {
             String line;
             while ((line = reader.readLine()) != null) {
                 String[] parts = line.split(",");
-                if (parts.length == 6) {
-                    cryptos.add(new CryptoCurrency(parts[0], Double.parseDouble(parts[1]),Double.parseDouble(parts[2]),Double.parseDouble(parts[3]),Double.parseDouble(parts[4]), Integer.parseInt(parts[5])));
+                if (parts.length == 7) {
+                    CryptoCurrency loadedCrypto = new CryptoCurrency(parts[0], Double.parseDouble(parts[1]),Double.parseDouble(parts[2]),Double.parseDouble(parts[3]),Double.parseDouble(parts[4]), Integer.parseInt(parts[5]));
+                    loadedCrypto.setInvestorsAmount(Integer.parseInt(parts[6]));
+                    cryptos.add(loadedCrypto);
                 }
             }
         }
+        if(cryptos.isEmpty()){ throw new NoSuchElementException("Nenhuma criptomoeda encontrada"); }
         return cryptos;
     }
+
 
     public List<String> loadCryptosToString() throws IOException{
         List<CryptoCurrency> cryptos = loadCryptos();
@@ -41,24 +48,34 @@ public class CryptoRepository {
         return stringOut;
     }
 
-    public CryptoCurrency loadCryptoByName(String  cryptoName) throws IOException {
+    public CryptoCurrency loadCryptoByName(String cryptoName) throws IOException {
         CryptoCurrency crypto = null;
         try (BufferedReader reader = new BufferedReader(new FileReader(FILE_PATH))) {
             String line;
             while ((line = reader.readLine()) != null) {
                 String[] parts = line.split(",");
-                if (parts.length == 6 && parts[0].equals(cryptoName)) {
+                if (parts.length == 7 && parts[0].equalsIgnoreCase(cryptoName)) {
                     crypto = new CryptoCurrency(parts[0], Double.parseDouble(parts[1]),Double.parseDouble(parts[2]),Double.parseDouble(parts[3]),Double.parseDouble(parts[4]), Integer.parseInt(parts[5]));
+                    crypto.setInvestorsAmount(Integer.parseInt(parts[6]));
                 }
             }
         }
-        if (crypto == null) { throw new IllegalArgumentException("Criptomoeda não encontrada"); }
+        if (crypto == null) { throw new NoSuchElementException("Criptomoeda não encontrada"); }
         return crypto;
     }
 
-    public void deleteCryptoByName(String name) throws IOException {
+    public void deleteCryptoByName(String cryptoName) throws IOException {
+        if(!cryptoExists(cryptoName)) { throw new NoSuchElementException("Criptomoeda não encontrada"); }
         List<CryptoCurrency> cryptos = loadCryptos();
-        cryptos.removeIf(crypto -> crypto.getName().equals(name));
+        CryptoCurrency removedCrypto = null;
+        for(CryptoCurrency crypto: cryptos) {
+            if (crypto.getName().equalsIgnoreCase(cryptoName)) {
+                removedCrypto = crypto;
+                if(removedCrypto.getInvestorsAmount() > 0) { throw new IllegalArgumentException("A criptomoeda tem investidores ativos e não pode ser removida"); }
+                break;
+            }
+        }
+        cryptos.remove(removedCrypto);
         // Reescreve o arquivo com a lista atualizada
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(FILE_PATH))) {
             for (CryptoCurrency crypto : cryptos) {
@@ -67,21 +84,35 @@ public class CryptoRepository {
         }
     }
 
-    public void updateCrypto(CryptoCurrency crypto) throws IOException {
-        List<CryptoCurrency> cryptos = loadCryptos();
-        boolean found = false;
-        for (CryptoCurrency cryptocurrency : cryptos) {
-            if (cryptocurrency.getName().equals(crypto.getName())) {
-                found = true;
-                break;
+    public void updateCrypto(CryptoCurrency updatedCrypto) throws IOException {
+        if(updatedCrypto == null || updatedCrypto.getName() == null ) { throw new IllegalArgumentException("Criptomoeda inválida");}
+        if(!cryptoExists(updatedCrypto.getName())) { throw new NoSuchElementException("Criptomoeda não encontrada"); }
+        List<CryptoCurrency> allCryptos = loadCryptos();
+        try (BufferedWriter writer = new BufferedWriter(new FileWriter(FILE_PATH))) {
+            for (CryptoCurrency crypto : allCryptos) {
+                if(crypto.getName().equalsIgnoreCase(updatedCrypto.getName())){
+                    writer.write(updatedCrypto.toString());
+                    writer.newLine();
+                }
+                else {
+                    writer.write(crypto.toString());
+                    writer.newLine();
+                }
             }
         }
-        if(found){
-            deleteCryptoByName(crypto.getName());
-            saveCrypto(crypto);
+    }
+
+    private boolean cryptoExists(String cryptoName) throws IOException {
+        try (BufferedReader reader = new BufferedReader(new FileReader(FILE_PATH))) {
+            String line;
+            while ((line = reader.readLine()) != null) {
+                String[] parts = line.split(",");
+                if (parts.length == 7 && parts[0].equalsIgnoreCase(cryptoName)) {
+                    return true;
+                }
+            }
         }
-        else{
-            throw new IllegalArgumentException();
-        }
+        return false;
     }
 }
+
